Extracted NpmRange.Includes method to interface IRange and updated usages
